Initial commit for project continuation
This commit marks the beginning of our continued development on the project, originally initiated by the TPEO student group at the University of Texas. Special thanks to @SarveshSub, @nayeonkim511, and @mercutio-22 for their foundational work.

We will carry forward their efforts to further develop and enhance the project, with the eventual goal of deploying the production version to https://library.austintexas.gov/events.

Credit:
- @SarveshSub
- @nayeonkim511
- @mercutio-22